+++ apps/api/src/services/queues/statistics-gathering/interfaces/interfaces.ts
@@ -1,0 +1,7 @@
+export interface BetterStackUptimeSlaResponse {
+  data: {
+    attributes: {
+      availability: number;
+    };
+  };
+}