+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added a _Storybook_ story for the copy-to-clipboard functionality of the value component
+
 ### Changed
 
 - Improved the language localization for German (`de`)