+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added the support to get the _Fear & Greed Index_ (market mood) from the `GHOSTFOLIO` data provider for self-hosted environments
+- Added support for the _Fear & Greed Index_ (market mood) via the `GHOSTFOLIO` data provider in self-hosted environments
 - Added a _Storybook_ story for the copy-to-clipboard functionality in the value component
 
 ### Changed