+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Improved the copy-to-clipboard functionality in the value component by providing a visual confirmation
 - Improved the language localization for German (`de`)
 - Upgraded `stripe` from version `22.2.3` to `22.3.2`